+++ src/modules/confirmation/ModuleConfirmation.vue
@@ -149,6 +149,9 @@ export default {
       _self.tx = tx;
       _self.tx.transactionFee = this.txFee;
       _self.resolver = resolver;
+      // this.txFee is not updating before reaching about here
+      _self.tx.transactionFee = this.txFee;
+      tx.transactionFee = this.txFee;
       _self.instance
         .signTransaction(tx)
         .then(res => {